+++ src/baseModule/index.ts
@@ -1,7 +1,8 @@
+/* eslint-disable check-file/filename-naming-convention */
 import { Module } from '../types/Module.js';
 import { ZumitoFramework } from '../ZumitoFramework.js';
-import { InteractionCreate } from './events/discord/interactionCreate.js';
-import { MessageCreate } from './events/discord/messageCreate.js';
+import { InteractionCreate } from './events/discord/InteractionCreate.js';
+import { MessageCreate } from './events/discord/MessageCreate.js';
 import { Guild } from './models/Guild.js';
 
 export class baseModule extends Module {
